% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momweight.R
\name{momweight}
\alias{momweight}
\title{Calculate weights for ...}
\usage{
momweight(X, S0, n)
}
\arguments{
\item{X}{vector of strikes}

\item{S0}{stock price}

\item{n}{power of the option contract
Note that, as there only a discrete number of options employed in this
approximation, the accuracy of the result will quickly decrease with N.}
}
\value{
Vector of N (4) first standardized moments
}
\description{
\code{momweight} ...
}
