% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option2stat.R
\name{Option2stat}
\alias{Option2stat}
\title{Compute implied standardized moments from calls and puts}
\usage{
Option2stat(XC, C, XP, P, S0, df, N)
}
\arguments{
\item{XC}{vector of call strikes}

\item{C}{vector of call prices}

\item{XP}{vector of put strikes}

\item{P}{vector of put prices}

\item{S0}{optional, if not given will be approximated along the way}

\item{df}{optional, if not given, will be approximated along the way}

\item{N}{optional, will return the N (4) first standardized moments.
Note that, as there only a discrete number of options employed in this
approximation, the accuracy of the result will quickly decrease with N.}
}
\value{
Vector of N (4) first standardized moments
}
\description{
\code{Option2stat} returns first N standardized moments of an asset's risk neutral log
return distribution from traded put and call options traded on that
asset, with equal time to maturity. Note that, as there only a discrete
number of options employed in this approximation, the accuracy of the
result will quickly decrease with N.
}
\examples{
Option2stat(DAX$XC,DAX$C,DAX$XP,DAX$P)

}
