% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option2price.R
\name{Options2price}
\alias{Options2price}
\title{Compute risk-neutral prices of a contract paying an asset's return or powers thereof from traded options.}
\usage{
Options2price(XC, C, XP, P, S0, df, N)
}
\arguments{
\item{XC}{vector of call strikes}

\item{C}{vector of call prices}

\item{XP}{vector of put strikes}

\item{P}{vector of put prices}

\item{S0}{optional, if not given will be approximated along the way}

\item{df}{optional, if not given, will be approximated along the way}

\item{N}{optional, will return the N (4) first standardized moments.
Note that, as there only a discrete number of options employed in this
approximation, the accuracy of the result will quickly decrease with N.}
}
\value{
Vector of N (4) first standardized moments
}
\description{
\code{Option2price} return the prices of contracts paying the first
N powers of the asset's future return $R^N=\log(ST/S0)^N$.
}
\examples{

}
